Add additional Provider support
By default, only 3 providers are available. This PR adds the ability for
a user to set additional Git providers with the global Git config.

The Git config structure includes a base URL as an argument, and commit
prefix and path prefix keys and values.

  [open "https://git.mydomain.dev"]
    commitprefix = commit
    pathprefix = tree

It can be set by editing the above into a .gitconfig or with the Git
CLI.

  git config --global open.https://git.mydomain.dev.commitprefix commit
  git config --global open.https://git.mydomain.dev.pathprefix tree

The resulting additional providers will be merged with the default
providers and work with `git open` calls.